fixedsize_stack uses malloc()
